making the _calc_mat_vec_wf fct in conftest to use among all test files needing it
